+++ src/components/mod.rs
@@ -10,6 +10,7 @@ mod create_branch;
 mod cred;
 mod diff;
 mod externaleditor;
+mod file_find;
 mod filetree;
 mod help;
 mod inspect_commit;
@@ -21,7 +22,6 @@ mod push_tags;
 mod rename_branch;
 mod reset;
 mod revision_files;
-mod revision_files_find;
 mod revision_files_popup;
 mod stashmsg;
 mod syntax_text;
